refactor(reservations): split reserve-bike command flow
