+++ apps/server/src/domain/reservations/index.ts
@@ -5,6 +5,7 @@ export * from "./repository/reservation-command.repository";
 export * from "./repository/reservation-query.repository";
 export * from "./services/cancel-reservation.service";
 export * from "./services/confirm-reservation.service";
+export * from "./services/commands/reserve-bike.service";
 export * from "./services/fixed-slot-template.service";
 export * from "./services/fixed-slot/fixed-slot.helpers";
 export * from "./services/fixed-slot/fixed-slot.service";
@@ -12,5 +13,4 @@ export type * from "./services/fixed-slot/fixed-slot.types";
 export * from "./services/reservation-command.service";
 export * from "./services/reservation-query.service";
 export * from "./services/reservation-stats.service";
-export * from "./services/reserve-bike.service";
 export * from "./types";
