--- apps/server/src/domain/reservations/services/reserve-bike.service.ts
@@ -1,346 +1,0 @@
-import { JobTypes } from "@mebike/shared/contracts/server/jobs";
-import { Effect, Option } from "effect";
-
-import type { BikeRepository } from "@/domain/bikes";
-import type { SubscriptionNotFound, SubscriptionNotUsable, SubscriptionUsageExceeded } from "@/domain/subscriptions/domain-errors";
-import type { WalletBalanceConstraint } from "@/domain/wallets/domain-errors";
-import type { DecreaseBalanceInput } from "@/domain/wallets/models";
-import type { ReservationOption } from "generated/prisma/client";
-
-import { env } from "@/config/env";
-import { makeBikeRepository } from "@/domain/bikes";
-import { getReservationFeeMinor, makePricingPolicyRepository } from "@/domain/pricing";
-import {
-  defectOn,
-  makeOvernightOperationsClosedError,
-  isWithinOvernightOperationsWindow,
-} from "@/domain/shared";
-import { toPrismaDecimal } from "@/domain/shared/decimal";
-import { makeStationQueryRepository } from "@/domain/stations";
-import { SubscriptionCommandServiceTag } from "@/domain/subscriptions";
-import { makeUserQueryRepository } from "@/domain/users";
-import { makeWalletRepository } from "@/domain/wallets";
-import { InsufficientWalletBalance, WalletNotFound } from "@/domain/wallets/domain-errors";
-import { enqueueOutboxJobInTx } from "@/infrastructure/jobs/outbox-enqueue";
-import { Prisma } from "@/infrastructure/prisma";
-import { PrismaTransactionError, runPrismaTransaction } from "@/lib/effect/prisma-tx";
-import { buildReservationConfirmedEmail } from "@/lib/email-templates";
-
-import type { ReservationServiceFailure } from "../domain-errors";
-import type { ReservationRow } from "../models";
-
-import {
-  ActiveReservationExists,
-  BikeAlreadyReserved,
-  BikeNotAvailable,
-  BikeNotFound,
-  BikeNotFoundInStation,
-  ReservationOptionNotSupported,
-  StationReservationAvailabilityTooLow,
-  SubscriptionRequired,
-} from "../domain-errors";
-import {
-  lockStationForReservationCheck,
-  requiredAvailableBikesForReservation,
-  stationCanAcceptReservation,
-} from "./reservation-availability-rule";
-import { ReservationCommandServiceTag } from "./reservation-command.service";
-import { ReservationQueryServiceTag } from "./reservation-query.service";
-
-export type ReserveBikeInput = {
-  readonly userId: string;
-  readonly bikeId: string;
-  readonly stationId: string;
-  readonly startTime: Date;
-  readonly reservationOption: ReservationOption;
-  readonly subscriptionId?: string | null;
-  readonly endTime?: Date | null;
-  readonly now?: Date;
-};
-
-export type ReserveBikeFailure
-  = | ReservationServiceFailure
-    | SubscriptionNotFound
-    | SubscriptionNotUsable
-    | SubscriptionUsageExceeded
-    | WalletNotFound
-    | InsufficientWalletBalance;
-
-const HOLD_MINUTES = env.RESERVATION_HOLD_MINUTES;
-const RESERVATION_TIME_ZONE = "Asia/Ho_Chi_Minh";
-
-function computeEndTime(startTime: Date, holdMinutes = HOLD_MINUTES): Date {
-  return new Date(startTime.getTime() + holdMinutes * 60 * 1000);
-}
-
-function formatReservationDateTime(value: Date): string {
-  return value.toLocaleString("vi-VN", { timeZone: RESERVATION_TIME_ZONE });
-}
-
-/**
- * Tao reservation hold cho user va reserve bike trong cung transaction.
- *
- * Thu tu xu ly chinh:
- * - Chan user hoac bike da co hold dang active.
- * - Validate bike thuoc dung tram va dang AVAILABLE.
- * - Khoa row Station de serialize availability check theo rule 50%.
- * - Thu prepaid qua wallet hoac tru usage subscription.
- * - Tao reservation, cap nhat bike, enqueue outbox jobs va email xac nhan.
- *
- * @param input Du lieu reservation can tao.
- * @returns Effect tra ve reservation PENDING neu thanh cong.
- */
-export function reserveBike(
-  input: ReserveBikeInput,
-): Effect.Effect<
-  ReservationRow,
-  ReserveBikeFailure,
-  | Prisma
-  | ReservationQueryServiceTag
-  | ReservationCommandServiceTag
-  | BikeRepository
-  | SubscriptionCommandServiceTag
-> {
-  return Effect.gen(function* () {
-    const { client } = yield* Prisma;
-    const reservationQueryService = yield* ReservationQueryServiceTag;
-    const reservationCommandService = yield* ReservationCommandServiceTag;
-    const subscriptionCommandService = yield* SubscriptionCommandServiceTag;
-    const now = input.now ?? new Date();
-
-    if (isWithinOvernightOperationsWindow(now)) {
-      return yield* Effect.fail(makeOvernightOperationsClosedError(now));
-    }
-
-    const reservation = yield* runPrismaTransaction(client, tx =>
-      Effect.gen(function* () {
-        const bikeRepo = makeBikeRepository(tx);
-        if (input.reservationOption === "FIXED_SLOT") {
-          return yield* Effect.fail(
-            new ReservationOptionNotSupported({ option: input.reservationOption }),
-          );
-        }
-
-        const existingByUser = yield* reservationQueryService.getCurrentHoldForUserNowInTx(
-          tx,
-          input.userId,
-          now,
-        );
-        if (Option.isSome(existingByUser)) {
-          return yield* Effect.fail(new ActiveReservationExists({ userId: input.userId }));
-        }
-
-        const activeReservation = yield* reservationQueryService.getLatestPendingOrActiveForUserInTx(
-          tx,
-          input.userId,
-        );
-        if (Option.isSome(activeReservation)) {
-          return yield* Effect.fail(new ActiveReservationExists({ userId: input.userId }));
-        }
-
-        const existingByBike = yield* reservationQueryService.getCurrentHoldForBikeNowInTx(
-          tx,
-          input.bikeId,
-          now,
-        );
-        if (Option.isSome(existingByBike)) {
-          return yield* Effect.fail(new BikeAlreadyReserved({ bikeId: input.bikeId }));
-        }
-
-        const bikeOpt = yield* bikeRepo.getById(input.bikeId);
-        if (Option.isNone(bikeOpt)) {
-          return yield* Effect.fail(new BikeNotFound({ bikeId: input.bikeId }));
-        }
-        const bike = bikeOpt.value;
-
-        if (!bike.stationId || bike.stationId !== input.stationId) {
-          return yield* Effect.fail(new BikeNotFoundInStation({
-            bikeId: input.bikeId,
-            stationId: input.stationId,
-          }));
-        }
-
-        if (bike.status !== "AVAILABLE") {
-          return yield* Effect.fail(new BikeNotAvailable({
-            bikeId: input.bikeId,
-            status: bike.status,
-          }));
-        }
-
-        const txStationRepo = makeStationQueryRepository(tx);
-        const stationOpt = yield* txStationRepo.getById(input.stationId);
-        if (Option.isNone(stationOpt)) {
-          return yield* Effect.die(new Error(
-            `Invariant violated: bike ${input.bikeId} references missing station ${input.stationId}`,
-          ));
-        }
-
-        yield* lockStationForReservationCheck(tx, input.stationId);
-
-        const availableBikes = yield* bikeRepo.countAvailableByStation(input.stationId);
-        const requiredAvailableBikes = requiredAvailableBikesForReservation(
-          stationOpt.value.totalCapacity,
-        );
-
-        if (!stationCanAcceptReservation({
-          totalCapacity: stationOpt.value.totalCapacity,
-          availableBikes,
-        })) {
-          return yield* Effect.fail(new StationReservationAvailabilityTooLow({
-            stationId: input.stationId,
-            totalCapacity: stationOpt.value.totalCapacity,
-            availableBikes,
-            requiredAvailableBikes,
-          }));
-        }
-
-        const subscriptionId: string | null = input.subscriptionId ?? null;
-        const pricingPolicy = yield* makePricingPolicyRepository(tx).getActive().pipe(
-          Effect.catchTag("ActivePricingPolicyNotFound", err => Effect.die(err)),
-          Effect.catchTag("ActivePricingPolicyAmbiguous", err => Effect.die(err)),
-        );
-
-        let prepaidMinor = getReservationFeeMinor(pricingPolicy);
-        let prepaid = toPrismaDecimal(prepaidMinor.toString());
-
-        if (input.reservationOption === "SUBSCRIPTION") {
-          if (!subscriptionId) {
-            return yield* Effect.fail(new SubscriptionRequired({ userId: input.userId }));
-          }
-          yield* subscriptionCommandService.useOne(tx, {
-            subscriptionId,
-            userId: input.userId,
-            now,
-          });
-
-          prepaid = toPrismaDecimal("0");
-          prepaidMinor = 0n;
-        }
-        else {
-          yield* debitWallet(makeWalletRepository(tx), {
-            userId: input.userId,
-            amount: prepaidMinor,
-            description: `Reservation prepaid ${input.userId}`,
-          });
-        }
-
-        const endTime = input.endTime ?? computeEndTime(input.startTime);
-
-        const reservation = yield* reservationCommandService.reserveHoldInTx(tx, {
-          userId: input.userId,
-          bikeId: input.bikeId,
-          stationId: input.stationId,
-          reservationOption: input.reservationOption,
-          subscriptionId,
-          startTime: input.startTime,
-          endTime,
-          prepaid,
-          pricingPolicyId: pricingPolicy.id,
-        });
-
-        const bikeReserved = yield* bikeRepo.reserveBikeIfAvailable(input.bikeId, now);
-        if (!bikeReserved) {
-          return yield* Effect.fail(new BikeAlreadyReserved({ bikeId: input.bikeId }));
-        }
-
-        if (reservation.endTime) {
-          const notifyAtMs = reservation.endTime.getTime()
-            - env.EXPIRY_NOTIFY_MINUTES * 60 * 1000;
-          const notifyAt = new Date(Math.max(now.getTime(), notifyAtMs));
-          const expireAt = new Date(Math.max(now.getTime(), reservation.endTime.getTime()));
-
-          yield* enqueueOutboxJobInTx(tx, {
-            type: JobTypes.ReservationNotifyNearExpiry,
-            payload: {
-              version: 1,
-              reservationId: reservation.id,
-            },
-            runAt: notifyAt,
-            dedupeKey: `reservation:notify:${reservation.id}`,
-          });
-
-          yield* enqueueOutboxJobInTx(tx, {
-            type: JobTypes.ReservationExpireHold,
-            payload: {
-              version: 1,
-              reservationId: reservation.id,
-            },
-            runAt: expireAt,
-            dedupeKey: `reservation:expire:${reservation.id}`,
-          });
-        }
-
-        // Confirmation email uses the older "success-reservation" template name.
-        // TODO(iot): send reservation "reserve" command once IoT integration is ready.
-        {
-          const txUserRepo = makeUserQueryRepository(tx);
-          const [userOpt, stationOpt] = yield* Effect.all([
-            txUserRepo.findById(reservation.userId),
-            txStationRepo.getById(reservation.stationId),
-          ]);
-
-          if (Option.isNone(userOpt)) {
-            return yield* Effect.die(new Error(
-              `Invariant violated: reservation ${reservation.id} references missing user ${reservation.userId}`,
-            ));
-          }
-          if (Option.isNone(stationOpt)) {
-            return yield* Effect.die(new Error(
-              `Invariant violated: reservation ${reservation.id} references missing station ${reservation.stationId}`,
-            ));
-          }
-
-          const email = buildReservationConfirmedEmail({
-            fullName: userOpt.value.fullname,
-            stationName: stationOpt.value.name,
-            bikeId: reservation.bikeId ?? input.bikeId,
-            startTimeLabel: formatReservationDateTime(reservation.startTime),
-            endTimeLabel: formatReservationDateTime(endTime),
-          });
-
-          yield* enqueueOutboxJobInTx(tx, {
-            type: JobTypes.EmailSend,
-            payload: {
-              version: 1,
-              to: userOpt.value.email,
-              kind: "raw",
-              subject: email.subject,
-              html: email.html,
-            },
-            runAt: now,
-            dedupeKey: `reservation-confirm:${reservation.id}`,
-          });
-        }
-
-        return reservation;
-      })).pipe(
-      defectOn(PrismaTransactionError),
-    );
-
-    return reservation;
-  });
-}
-
-/**
- * Tru tien vi cho prepaid reservation va map loi wallet ve domain error cua reservation flow.
- *
- * @param repo Wallet repository dang dung trong transaction.
- * @param input Du lieu tru tien.
- * @returns Effect thanh cong khi tru tien xong hoac fail voi wallet-related domain error.
- */
-function debitWallet(
-  repo: ReturnType<typeof makeWalletRepository>,
-  input: DecreaseBalanceInput,
-) {
-  return repo.decreaseBalance(input).pipe(
-    Effect.catchTag("WalletRecordNotFound", () =>
-      Effect.fail(new WalletNotFound({ userId: input.userId }))),
-    Effect.catchTag("WalletBalanceConstraint", (err: WalletBalanceConstraint) =>
-      Effect.fail(new InsufficientWalletBalance({
-        walletId: err.walletId,
-        userId: err.userId,
-        balance: err.balance,
-        attemptedDebit: err.attemptedDebit,
-      }))),
-  );
-}
